STLs for the ZuluSCSI Wide SCA rev 2026+ board
In the directory `ZuluSCSI-Wide-SCA-pcb-direct [rev 2026a+]`
contains the FreeCAD file `Brackets-and-Shroud.FCStd` which is the
source for STLs:
`Bracket_Left.stl`
`Bracket_Right.stl`
`Connector-Shroud.stl`.

The files are for the ZuluSCSI Wide SCA rev 2026a and future boards
that connect directly to the SCA connector using the ZuluSCSI's pcb
edge. The Connector shroud does a pretty good job at avoiding inserting
the ZuluSCSI upside down.

While the brackets has the correct height alignment for the PCB edge
SCA connector and holes for mounting in a hard drive bay.

As of now the 1/4" 440 screw is a bit to long for the SD/USB side
of the bracket PCB mounts.
A pair of washers would probably fix the issue.

The same screws may be a little to short for the shroud SCA side of
the board. A 3/8" 440 screw is a bit too long, but the shroud mounting
holes could be raised to accommodate their length.